Generation 2
A comprehensive update, overall, added several new quality of life features, as well as some descent fall backs for now. In this update to my past time game, I added functionality for a timer that starts when you select a difficulty. It will track the time you spend on each puzzle.

Also added a modal that will appear when you finish a puzzle, with a few different messages depending on how you did. It will also allow for a new puzzle at the same difficulty, or bring you back to the menu. (In case you would rather switch the difficulty, or look at my atm rather bleak welcome screen.)

I also added some rather helpful quality of life features, that will highlight every square that has been filled of the same value. It can be activate with hovering over an element, or focusing it. If you click on a solved square, it will also lock the styles for the squares of the same value.

And lastly, ive increased the styles across the board, for better visuals, as well as adding the unique styles for the modal.

In the next update, it will mainly be a style overhaul.

Depending on how long it takes till i am satisfied with styles, i may or may or may include a puzzle validator. The plan is to overhaul the boardGenerator, to run another function replacing is safe, to validate not only the current value, but the whole board to make sure there is only one solution. It will also ensure that the puzzle is solvable solely on logic, and lastly, it will feed values, that i will use to validate the inputs, to remove my (hopefully) sole bug, where sometimes values that are the only logical solution, will sometimes evaluate as false.

Wow that was a wrist-full! Get it? But  really, I may be developing carpel tunnel. J.K.... maybe not, anyways if you read this you're awesome, goodbye, have a beautiful time!

Peace,
    theEphemeralDev